修正1　destroy

+++ app/controllers/items_controller.rb
@@ -1,5 +1,5 @@
 class ItemsController < ApplicationController
-  before_action :authenticate_user!, except: [:index, :show, :destroy]
+  before_action :authenticate_user!, except: [:index, :show]
   before_action :set_item, only: [:edit, :show, :update, :edit, :destroy]
 
   def index
@@ -20,6 +20,7 @@ class ItemsController < ApplicationController
   end
 
   def destroy
+    redirect_to root_path unless current_user.id == @item.user_id
    if @item.destroy
      redirect_to root_path
     else
